requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.3.2
Django==1.6
enum34==1.1.6
gunicorn==19.9.0
ipython==5.8.0
ipython-genutils==0.2.0
netaddr==0.7.19
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pkg-resources==0.0.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pyrad==2.1
reportlab==3.5.13
scandir==1.9.0
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
